feat: move MessengerProducerContext to its own file
