+++ packages/talos_messenger_actions/src/kafka/context.rs
@@ -1,0 +1,42 @@
+use futures_executor::block_on;
+use log::info;
+use rdkafka::{producer::ProducerContext, ClientContext, Message};
+use talos_messenger_core::core::MessengerChannelFeedback;
+use tokio::sync::mpsc;
+
+#[derive(Debug)]
+pub struct MessengerProducerDeliveryOpaque {
+    pub version: u64,
+    pub total_publish_count: u32,
+}
+
+pub struct MessengerProducerContext {
+    pub tx_feedback_channel: mpsc::Sender<MessengerChannelFeedback>,
+}
+
+impl ClientContext for MessengerProducerContext {}
+impl ProducerContext for MessengerProducerContext {
+    type DeliveryOpaque = Box<MessengerProducerDeliveryOpaque>;
+
+    fn delivery(&self, delivery_result: &rdkafka::producer::DeliveryResult<'_>, delivery_opaque: Self::DeliveryOpaque) {
+        let result = delivery_result.as_ref();
+
+        let version = delivery_opaque.version;
+
+        match result {
+            Ok(msg) => {
+                info!("Message {:?} {:?}", msg.key(), msg.offset());
+                // TODO: GK - what to do on error? Panic?
+                let _ = block_on(self.tx_feedback_channel.send(MessengerChannelFeedback::Success(
+                    version,
+                    "kafka".to_string(),
+                    delivery_opaque.total_publish_count,
+                )));
+            }
+            Err(err) => {
+                // TODO: GK - what to do on error? Panic?
+                let _ = block_on(self.tx_feedback_channel.send(MessengerChannelFeedback::Error(version, err.0.to_string())));
+            }
+        }
+    }
+}
