+++ packages/talos_messenger_actions/src/kafka/mod.rs
@@ -1,3 +1,4 @@
+pub mod context;
 pub mod models;
 pub mod producer;
 pub mod service;
